requirements.txt
flask
vectorizer
openai
python-dotenv

